LayoutTests/media/controls/statusDisplay-expected.txt

This tests for the content in the status display.

This test only runs in DRT!


EVENT: canplaythrough
PASS: We are using the apple idiom

Get the status label state

PASS: Name is 'Status Display'
PASS: Has hidden class
PASS: Content is Empty
PASS: CSS display: none

Testing finished.
